refactor(product-category-crud): Separate controller and views to Admin
