+++ resources/views/admin/product-category/create.blade.php
